added error handling; modified controllers

+++ backend/src/controllers/user.controller.ts
@@ -1,9 +1,19 @@
+import {inject} from '@loopback/core';
 import {repository} from '@loopback/repository';
-import {del, getModelSchemaRef, param, post, requestBody} from '@loopback/rest';
+import {
+  del,
+  getModelSchemaRef,
+  HttpErrors,
+  param,
+  post,
+  requestBody,
+  RequestContext,
+} from '@loopback/rest';
 import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
 import {User} from '../models';
 import {UserRepository} from '../repositories';
+import {checkAuth} from '../utils';
 
 class UserCredentials {
   username: string;
@@ -12,10 +22,11 @@ class UserCredentials {
 
 export class UserController {
   constructor(
-    @repository(UserRepository)
-    public userRepository: UserRepository,
+    @inject.context() public context: RequestContext,
+    @repository(UserRepository) public userRepository: UserRepository,
   ) {}
 
+  // SIGNUP
   @post('/signup', {
     responses: {
       '200': {
@@ -35,23 +46,27 @@ export class UserController {
       },
     })
     userCredentials: UserCredentials,
-  ): Promise<User> {
+  ): Promise<{token: string}> {
     const foundUser = await this.userRepository.findOne({
       where: {username: userCredentials.username},
     });
     if (foundUser) {
-      throw {
-        code: 409,
-        message: 'Username already exists',
-      };
+      throw new HttpErrors.Conflict('Username already exists');
     }
     const password_hash = await bcrypt.hash(userCredentials.password, 10);
-    const newUser = new User();
+    let newUser = new User();
     newUser.username = userCredentials.username;
     newUser.password_hash = password_hash;
-    return this.userRepository.create(newUser);
+    newUser = await this.userRepository.create(newUser);
+    const token = jwt.sign(
+      {id: newUser.id, username: newUser.username},
+      'JWT_secret',
+      {expiresIn: '2h'},
+    );
+    return {token: token};
   }
 
+  // LOGIN
   @post('/login', {
     responses: {
       '200': {
@@ -100,12 +115,10 @@ export class UserController {
         }
       }
     } catch (error) {}
-    throw {
-      code: 401,
-      message: 'Username or password is incorrect',
-    };
+    throw new HttpErrors.Unauthorized('Failed to log in');
   }
 
+  // DELETE
   @del('/users/{id}', {
     responses: {
       '204': {
@@ -114,6 +127,10 @@ export class UserController {
     },
   })
   async deleteById(@param.path.number('id') id: number): Promise<void> {
+    const currentUser = checkAuth(this.context);
+    if (currentUser.id !== id) {
+      throw new HttpErrors.Forbidden('Not allowed');
+    }
     await this.userRepository.deleteById(id);
   }
 }
